Reorganize node tests
