+++ src/node/expression.rs
@@ -60,3 +60,96 @@ impl<'a> Iterator for Iter<'a> {
         self.iterator.next()
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn new_expressions() {
+        let t1: Term = Term::Terminal(String::from("terminal"));
+        let nt1: Term = Term::Nonterminal(String::from("nonterminal"));
+        let t2: Term = Term::Terminal(String::from("terminal"));
+        let nt2: Term = Term::Nonterminal(String::from("nonterminal"));
+
+        let e1: Expression = Expression::from_parts(vec![nt1, t1]);
+        let mut e2: Expression = Expression::new();
+        e2.add_term(nt2);
+        e2.add_term(t2);
+
+        assert_eq!(e1, e2);
+    }
+
+    #[test]
+    fn add_term_to_expression() {
+        let mut terms = vec![
+            Term::Terminal(String::from("A")),
+            Term::Terminal(String::from("C")),
+            Term::Terminal(String::from("G")),
+        ];
+
+        let mut dna_expression = Expression::from_parts(terms.clone());
+        assert_eq!(dna_expression.terms_iter().count(), terms.len());
+
+        // oops forgot "T"
+        let forgotten = Term::Terminal(String::from("T"));
+        dna_expression.add_term(forgotten.clone());
+        terms.push(forgotten);
+        assert_eq!(dna_expression.terms_iter().count(), terms.len());
+
+        // check all terms are there
+        for term in dna_expression.terms_iter() {
+            assert!(terms.contains(term), "{} was not in terms", term);
+        }
+    }
+
+    #[test]
+    fn remove_term_from_expression() {
+        let terms = vec![
+            Term::Terminal(String::from("A")),
+            Term::Terminal(String::from("C")),
+            Term::Terminal(String::from("G")),
+            Term::Terminal(String::from("T")),
+            Term::Terminal(String::from("Z")),
+        ];
+
+        let mut dna_expression = Expression::from_parts(terms.clone());
+        assert_eq!(dna_expression.terms_iter().count(), terms.len());
+
+        // oops "Z" isn't a dna base
+        let accident = Term::Terminal(String::from("Z"));
+        let removed = dna_expression.remove_term(&accident);
+
+        // the removed element should be the accident
+        assert_eq!(Some(accident.clone()), removed);
+        // number of terms should have decreased
+        assert_eq!(dna_expression.terms_iter().count(), terms.len() - 1);
+        // the accident should no longer be found in the terms
+        assert_eq!(
+            dna_expression.terms_iter().find(|&term| *term == accident),
+            None
+        );
+    }
+
+    #[test]
+    fn remove_nonexistent_term_from_expression() {
+        let terms = vec![
+            Term::Terminal(String::from("A")),
+            Term::Terminal(String::from("C")),
+            Term::Terminal(String::from("G")),
+            Term::Terminal(String::from("T")),
+        ];
+
+        let mut dna_expression = Expression::from_parts(terms.clone());
+        assert_eq!(dna_expression.terms_iter().count(), terms.len());
+
+        // oops "Z" isn't a dna base
+        let nonexistent = Term::Terminal(String::from("Z"));
+        let removed = dna_expression.remove_term(&nonexistent);
+
+        // the removed element should be the accident
+        assert_eq!(None, removed);
+        // number of terms should not have decreased
+        assert_eq!(dna_expression.terms_iter().count(), terms.len());
+    }
+}
