+++ tests/iterate.rs
@@ -1,0 +1,69 @@
+extern crate bnf;
+use bnf::node::Term;
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn iterate_grammar() {
+        let dna_productions = "
+            <dna> ::= <dna> | <base> <dna>;
+            <base> ::= \"A\" | \"C\" | \"G\" | \"T\";";
+
+        let dna_grammar = bnf::parse(dna_productions);
+
+        let left_hand_terms: Vec<&Term> = dna_grammar
+            .productions_iter()
+            .map(|ref prod| &prod.lhs)
+            .collect();
+
+        // should be as many left hand terms as productions
+        assert_eq!(
+            dna_grammar.productions_iter().count(),
+            left_hand_terms.len()
+        );
+
+        let right_hand_terms: Vec<&Term> = dna_grammar
+            .productions_iter()
+            .flat_map(|prod| prod.rhs_iter())
+            .flat_map(|expr| expr.terms_iter())
+            .collect();
+
+        // check nonterminals are in left and right hand terms collection
+        for term in ["dna", "base"].into_iter() {
+            let term_string = String::from(*term);
+            let expected_nonterminal = Term::Nonterminal(term_string);
+
+            assert!(
+                left_hand_terms.contains(&&expected_nonterminal),
+                "{} was not in left hand terms",
+                expected_nonterminal
+            );
+
+            assert!(
+                right_hand_terms.contains(&&expected_nonterminal),
+                "{} was not in right hand terms",
+                expected_nonterminal
+            );
+        }
+
+        // any term which appears on the right but not left is a terminal
+        let only_right_terms: Vec<&Term> = right_hand_terms
+            .into_iter()
+            .filter(|term| !left_hand_terms.contains(term))
+            .collect();
+
+        // check terminals are only on right hand side
+        for term in ["A", "C", "G", "T"].into_iter() {
+            let term_string = String::from(*term);
+            let expected_terminal = Term::Terminal(term_string);
+
+            assert!(
+                only_right_terms.contains(&&expected_terminal),
+                "{} was not in left hand terms",
+                expected_terminal
+            );
+        }
+    }
+}
